+++ src/include/wal.h
@@ -30,7 +30,7 @@ struct WAL *WAL_new(char *filename);
 
 int WAL_load_memtable(struct WAL *wal, struct MemTable *memtable);
 
-int WAL_append(struct WAL *wal, const char *key, size_t key_len, long long value_loc);
+int WAL_append(struct WAL *wal, const char *key, size_t key_len, int64_t value_loc);
 
 int WAL_sync(const struct WAL *wal);
